feat: device table --> homepage
