+++ src/routes/+layout.svelte
@@ -16,7 +16,6 @@
 <div>
 	<Navigation.Bar>
 		<Navigation.Tile label="Home" href="/"></Navigation.Tile>
-		<Navigation.Tile label="Device Table" href="/deviceTable"></Navigation.Tile>
 		<Navigation.Tile label="Add Device" href="/addDevice"></Navigation.Tile>
 	</Navigation.Bar>
 </div>
