+++ src/routes/[deviceID]/+page.ts
