+++ src/routes/[deviceID]/editDevice/+page.svelte
