+++ src/routes/[deviceID]/editDevice/+page.ts
